feat: 수정, 삭제 버튼에 같은 row 아이템 데이터 저장 및 사용

+++ src/js/controllers/manageItemController.ts
@@ -1,5 +1,6 @@
 import ManageItemView from '../views/mangeItemView';
 import VendingMachine from '../vendingMachine/vendingMachine';
+import { CONFIRM_MESSAGE } from '../constants/confirmConstants';
 import { CUSTOM_EVENT } from '../constants/appContants';
 import { checkDuplicatedItem, validateAddItemInput } from '../validates/validates';
 import { ItemType, TableItemChangeDetailType, TableItemDeleteDetailType } from '../types/types';
@@ -56,6 +57,8 @@ export default class ManageItemController {
 
   handleTableItemDelete(event: CustomEvent) {
     const { item }: TableItemDeleteDetailType = event.detail;
-    this.vendingMachine.deleteItem(item);
+    if (window.confirm(CONFIRM_MESSAGE.DELETE)) {
+      this.vendingMachine.deleteItem(item);
+    }
   }
 }
